+++ src/file_tree/sort_type.rs
@@ -1,5 +1,0 @@
-#[derive(Copy, Clone)]
-pub enum SortType {
-    Desc,
-    Asc
-}