+++ CHANGELOG.md
@@ -6,3 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [0.1.0] - 2021-05-08
 
 First release.
+
+## [1.0.0] - 2021-05-08
+
+Did a complete rewrite with emphasis on an intuitive interface and performance. Notable changes:
+- Respects `.gitignore`.
+- Parallel filesystem traversal.
+- Completely new CLI. `$ erdtree -h` for usage info.
+- Uses `LS_COLORS` environment variable for file coloring.